requirements.txt
asgiref==3.5.2
asttokens==2.0.8
attrs==22.1.0
autopep8==1.7.0
backcall==0.2.0
coverage==6.5.0
decorator==5.1.1
Django==4.1.1
django-debug-toolbar==3.7.0
django-decouple==2.1
executing==1.1.1
Faker==14.2.1
flake8==5.0.4
iniconfig==1.1.1
ipython==8.5.0
jedi==0.18.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mypy==0.971
mypy-extensions==0.4.3
packaging==21.3
parameterized==0.8.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pluggy==1.0.0
prompt-toolkit==3.0.31
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pycodestyle==2.9.1
pyflakes==2.5.0
Pygments==2.13.0
pyparsing==3.0.9
pytest==7.1.3
pytest-django==4.5.2
python-dateutil==2.8.2
python-dotenv==0.21.0
six==1.16.0
sqlparse==0.4.2
stack-data==0.5.1
toml==0.10.2
tomli==2.0.1
traitlets==5.4.0
typing_extensions==4.3.0
watchdog==2.1.9
wcwidth==0.2.5
whitenoise==6.2.0
